7.0.02
Fixed an error that could happen when sorting by faction.
Fixed overview not updating when switching between emissary quests.
Fixed overview not updating when selecting a quest on the world map.
Fixed an issue with quest details while the world quest tab was open.
Fixed taint issues with default raidframes.